# api/api.yaml
openapi: 3.0.3
info:
  title: Timeship API v2 (RESTful Edition)
  description: |
    RESTful file management API using unified node concept.
    
    All filesystem objects (files and directories) are treated as "nodes" with a type field.
    This mirrors the Unix/Linux filesystem philosophy where everything is an inode.
    
    Key concepts:
    - Nodes are identified by their path (compound key with slashes)
    - Node type (file/dir) is determined by the response, not the URL
    - Paths are hierarchical and natural - empty path = storage root
    - Operations are RESTful resources (moves, copies, archives)
    - Storage identifier is implicit in path prefix (e.g., "local://path")
  version: 2.0.0
servers:
  - url: /api/
    description: API v2 server
...
components:
  schemas:
    NodeType:
      type: string
      enum: [file, dir]
      description: Type of the filesystem node
    Node:
      type: object
      description: |
        Unified representation of any filesystem object (file or directory).
        Storage identifier is implicit in the path prefix (e.g., "local://path").
      required:
        - path
        - type
        - basename
        - extension
        - last_modified
        - file_size
      properties:
        path:
          type: string
          description: Full path with storage prefix
          example: 'local://documents/reports/2024/report.pdf'
        type:
          $ref: '#/components/schemas/NodeType'
        basename:
          type: string
          description: Base name of the node
          example: 'report.pdf'
        extension:
          type: string
          description: File extension (empty string for directories)
          example: 'pdf'
        mime_type:
          type: string
          description: MIME type (only present for files when detection succeeds)
          example: 'application/pdf'
        file_size:
          type: integer
          format: int64
          description: Size in bytes (0 for directories)
          example: 1048576
        last_modified:
          type: integer
          format: int64
          description: Unix timestamp of last modification
          example: 1698364800
        url:
          type: string
          nullable: true
          description: Public URL for the file (present when URL resolver is configured, null otherwise)
          example: 'https://cdn.example.com/documents/reports/2024/report.pdf'
        dir:
          type: string
          description: Parent directory path with storage prefix (only present in search results)
          example: 'local://documents/reports/2024'
    NodeList:
      type: object
      description: |
        Response containing list of nodes.
      required:
        - dirname
        - files
        - read_only
        - storages
      properties:
        dirname:
          type: string
          description: Current directory path with storage prefix
          example: 'local://documents/reports'
        files:
          type: array
          description: Child nodes in the current directory
          items:
            $ref: '#/components/schemas/Node'
        read_only:
          type: boolean
          description: Whether the current storage is read-only
          example: false
        storages:
          type: array
          description: Available storage identifiers
          items:
            type: string
          example: ['local', 's3', 'minio']
        total_size:
          type: integer
          format: int64
          description: |
            Total size in bytes of all files in this directory and subdirectories.
            Only included when requested via fields=(total_size) query parameter.
            Computed using parallel directory traversal for optimal performance.
          example: 104857600
    CreateNodeRequest:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
          pattern: '^[^\\/?%*:|"<>]+$'
          description: Name of the node to create
          example: 'new-folder'
        type:
          $ref: '#/components/schemas/NodeType'
        content:
          type: string
          description: Initial content (only for files)
    UpdateNodeRequest:
      type: object
      properties:
        name:
          type: string
          pattern: '^[^\\/?%*:|"<>]+$'
          description: New name for the node (rename)
        content:
          type: string
          description: Updated content (only for files)
...
    ErrorResponse:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: boolean
          enum: [false]
          description: Always false for error responses
          example: false
        message:
          type: string
          description: Human-readable error message
          example: 'Invalid node name.'
...
  parameters:
    storage:
      name: storage
      in: path
      required: true
      schema:
        type: string
      description: Storage backend identifier
      example: local
    nodePath:
      name: path...
      in: path
      required: true
      schema:
        type: string
      style: simple
      allowReserved: true
      description: |
        Node path (can contain slashes). Path is relative to storage root.
        Use empty string or single space for storage root.
        
        Note: OpenAPI path parameters cannot be truly optional, but empty string 
        is supported by the router. The spec uses '*' suffix to match everything 
        including empty paths.
      examples:
        root:
          value: " "
          summary: Storage root (use space in spec, empty in actual URLs)
        file:
          value: "documents/report.pdf"
          summary: File path
        directory:
          value: "documents/reports/2024"
          summary: Directory path
    getNodesType:
      name: type
      in: query
      schema:
        $ref: '#/components/schemas/NodeType'
      description: Filter children by type (for directories)
    getNodesFilter:
      name: filter
      in: query
      schema:
        type: string
      description: Filename pattern (glob-style, e.g., *.pdf)
      example: '*.pdf'
    getNodesSearch:
      name: search
      in: query
      schema:
        type: string
      description: Search query - searches recursively from this path
      example: 'report'
    getNodesChildren:
      name: children
      in: query
      schema:
        type: boolean
        default: true
      description: Include children in response (for directories)
    getNodesDownload:
      name: download
      in: query
      schema:
        type: boolean
        default: false
      description: Set Content-Disposition to attachment (for files)
    getNodesSort:
      name: sort
      in: query
      schema:
        type: string
        enum: [name, size, modified_at, type]
      description: Sort field for children
    getNodesOrder:
      name: order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
      description: Sort order
    getNodesFields:
      name: fields
      in: query
      schema:
        type: string
      description: |
        Comma-separated list of optional fields to include in the response.
        Each field must be wrapped in parentheses.
        
        Available fields:
        - (total_size): Include total size of directory and all subdirectories
        
        Example: fields=(total_size)
      example: '(total_size)'
    getNodesSnapshot:
      name: snapshot
      in: query
      schema:
        type: string
      description: |
        Snapshot identifier in format "type:backend-id" (e.g., "zfs:tank@daily-2024-10-28").
        When provided, returns the node as it existed in that snapshot.
      example: "zfs:tank@daily-2024-10-28"
    deleteNodesRecursive:
      name: recursive
      in: query
      schema:
        type: boolean
        default: true
      description: Delete recursively (for directories)
...
  responses:
    nodeSuccess200:
      description: Node information or content
      headers:
        Content-Type:
          schema:
            type: string
          description: Response content type
        Content-Disposition:
          schema:
            type: string
          description: Disposition header (for file downloads)
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/Node'
              - $ref: '#/components/schemas/NodeList'
          examples:
            directory:
              summary: Directory listing
              value:
                dirname: local://documents
                read_only: false
                storages: [local, s3]
                files:
                  - path: local://documents/reports
                    type: dir
                    basename: reports
                    extension: ""
                    last_modified: 1698364800
                    file_size: 0
                  - path: local://documents/readme.txt
                    type: file
                    basename: readme.txt
                    extension: txt
                    mime_type: text/plain
                    file_size: 1024
                    last_modified: 1698364800
                    url: https://cdn.example.com/documents/readme.txt
            file:
              summary: File metadata
              value:
                path: local://documents/report.pdf
                type: file
                basename: report.pdf
                extension: pdf
                mime_type: application/pdf
                file_size: 1048576
                last_modified: 1698364800
                url: https://cdn.example.com/documents/report.pdf
            search:
              summary: Search results
              value:
                dirname: local://documents
                read_only: false
                storages: [local, s3]
                files:
                  - path: local://documents/2024/report.pdf
                    type: file
                    basename: report.pdf
                    extension: pdf
                    mime_type: application/pdf
                    file_size: 2097152
                    last_modified: 1698364800
                    dir: local://documents/2024
                    url: https://cdn.example.com/documents/2024/report.pdf
                  - path: local://documents/reports/annual-report.pdf
                    type: file
                    basename: annual-report.pdf
                    extension: pdf
                    mime_type: application/pdf
                    file_size: 1536000
                    last_modified: 1698364800
                    dir: local://documents/reports
                    url: https://cdn.example.com/documents/reports/annual-report.pdf
        application/octet-stream:
          schema:
            type: string
            format: binary
        text/plain:
          schema:
            type: string
    nodeCreated201:
      description: Node created successfully
      headers:
        Location:
          schema:
            type: string
          description: URL to the created node
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Node'
    nodeNotFound404:
      description: Node not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    badRequest400:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    nodeConflict409:
      description: Node already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
paths:
...
  /storages/{storage}/nodes:
    parameters:
      - $ref: '#/components/parameters/storage'
      
    get:
      summary: Get storage root information or content
      description: |
        Get information about storage root, or its content if accessible as a file.
        This is a convenience endpoint for accessing the root without a path parameter.
        
        For storage roots: Returns root metadata and lists children.
        
        When ?snapshot= is provided, returns the node as it existed in that snapshot.
        
        Content negotiation:
        - Accept: application/json → Returns node metadata
        - Accept: application/octet-stream → Returns file content (binary)
        - Accept: text/* → Returns file content (text)
      tags: [Nodes]
      parameters:
        - $ref: '#/components/parameters/getNodesType'
        - $ref: '#/components/parameters/getNodesFilter'
        - $ref: '#/components/parameters/getNodesSearch'
        - $ref: '#/components/parameters/getNodesChildren'
        - $ref: '#/components/parameters/getNodesDownload'
        - $ref: '#/components/parameters/getNodesSort'
        - $ref: '#/components/parameters/getNodesOrder'
        - $ref: '#/components/parameters/getNodesFields'
        - $ref: '#/components/parameters/getNodesSnapshot'
      responses:
        '200':
          $ref: '#/components/responses/nodeSuccess200'
        '404':
          $ref: '#/components/responses/nodeNotFound404'
                
    post:
      summary: Create a new child node at storage root
      description: |
        Create a new file or directory at the storage root.
      tags: [Nodes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNodeRequest'
            examples:
              directory:
                summary: Create directory
                value:
                  name: "2024-reports"
                  type: dir
              file:
                summary: Create file with content
                value:
                  name: "notes.txt"
                  type: file
                  content: "Hello, world!"
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: File to upload
                name:
                  type: string
                  description: Optional custom filename (defaults to uploaded filename)
      responses:
        '201':
          $ref: '#/components/responses/nodeCreated201'
        '400':
          $ref: '#/components/responses/badRequest400'
        '409':
          $ref: '#/components/responses/nodeConflict409'
  /storages/{storage}/nodes/{path...}:
    parameters:
      - $ref: '#/components/parameters/storage'
      - $ref: '#/components/parameters/nodePath'
      
    get:
      summary: Get node information or content
      description: |
        Get information about a node, or its content if it's a file.
        Works for both storage root (empty path) and nested paths.
        
        For directories: Returns node metadata and lists children.
        For files: Returns file content (use Accept header for format).
        
        When ?snapshot= is provided, returns the node as it existed in that snapshot.
        
        Content negotiation:
        - Accept: application/json → Returns node metadata
        - Accept: application/octet-stream → Returns file content (binary)
        - Accept: text/* → Returns file content (text)
      tags: [Nodes]
      parameters:
        - $ref: '#/components/parameters/getNodesType'
        - $ref: '#/components/parameters/getNodesFilter'
        - $ref: '#/components/parameters/getNodesSearch'
        - $ref: '#/components/parameters/getNodesChildren'
        - $ref: '#/components/parameters/getNodesDownload'
        - $ref: '#/components/parameters/getNodesSort'
        - $ref: '#/components/parameters/getNodesOrder'
        - $ref: '#/components/parameters/getNodesFields'
        - $ref: '#/components/parameters/getNodesSnapshot'
      responses:
        '200':
          $ref: '#/components/responses/nodeSuccess200'
        '404':
          description: Node not found or snapshot not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
                
    post:
      summary: Create a new child node
      description: |
        Create a new file or directory as a child of this path.
        For storage root, creates at root level.
      tags: [Nodes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNodeRequest'
            examples:
              directory:
                summary: Create directory
                value:
                  name: "2024-reports"
                  type: dir
              file:
                summary: Create file with content
                value:
                  name: "notes.txt"
                  type: file
                  content: "Hello, world!"
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: File to upload
                name:
                  type: string
                  description: Optional custom filename (defaults to uploaded filename)
      responses:
        '201':
          $ref: '#/components/responses/nodeCreated201'
        '400':
          $ref: '#/components/responses/badRequest400'
        '409':
          $ref: '#/components/responses/nodeConflict409'
                
    patch:
      summary: Update node metadata or content
      description: |
        Update node name (rename) or content (for files).
        Partial updates are supported.
      tags: [Nodes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNodeRequest'
            examples:
              rename:
                summary: Rename node
                value:
                  name: "new-name.pdf"
              updateContent:
                summary: Update file content
                value:
                  content: "Updated content..."
      responses:
        '200':
          description: Node updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
        '404':
          description: Node not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
                
    delete:
      summary: Delete a node
      description: |
        Delete a file or directory.
        For directories, all children are deleted recursively by default.
      tags: [Nodes]
      parameters:
        - $ref: '#/components/parameters/deleteNodesRecursive'
      responses:
        '204':
          description: Node deleted successfully
        '404':
          $ref: '#/components/responses/nodeNotFound404'
        '409':
          description: Directory not empty and recursive=false
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...